renamed prompt buttons to modal buttons
