+++ src/components/ModalButtons/RegisterPromptButton.tsx
